refactor: 공지사항 인디케이터 로직 커스텀 훅으로 추출
