+++ src/pages/posts/list/PostList.tsx
@@ -4,48 +4,19 @@ import StudentCouncilListItem from "./components/StudentCouncilListItem";
 import { generalPostsDummy } from "./constant/GeneralPostsDummy";
 import { studentCouncilPostsDummy } from "./constant/StudentCouncilPostsDummy";
 import * as styles from "./PostList.css";
-import { useState, useRef, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
+import { useSlideIndicator } from "./hooks/useSlideIndicator";
 import { ROUTES } from "@router/constant/Routes";
 
 export default function PostList() {
   const navigator = useNavigate();
 
-  const [currentSlide, setCurrentSlide] = useState(0);
-  const containerRef = useRef<HTMLDivElement>(null);
-
   const totalSlides = studentCouncilPostsDummy.length;
-
-  const handleIndicatorClick = (index: number) => {
-    setCurrentSlide(index);
-    if (containerRef.current) {
-      const itemWidth = 29 + 16;
-      const scrollAmount = index * itemWidth;
-
-      containerRef.current.scrollTo({
-        left: scrollAmount,
-        behavior: "smooth",
-      });
-    }
-  };
-
-  // 스크롤 이벤트 감지
-  useEffect(() => {
-    const container = containerRef.current;
-    if (!container) return;
-
-    const handleScroll = () => {
-      const scrollLeft = container.scrollLeft;
-      const itemWidth = 29 + 20; // 아이템 너비 + gap
-
-      const currentIndex = Math.floor(scrollLeft / (totalSlides * itemWidth));
-
-      setCurrentSlide(Math.max(0, Math.min(currentIndex, totalSlides - 1)));
-    };
-
-    container.addEventListener("scroll", handleScroll);
-    return () => container.removeEventListener("scroll", handleScroll);
-  }, [totalSlides]);
+  const { currentSlide, containerRef, handleIndicatorClick } = useSlideIndicator({
+    totalSlides,
+    itemWidth: 29,
+    itemGap: 20,
+  });
 
   return (
     <div className={styles.container}>
